requirements.txt
# Project dependencies

Authlib
Django
Django-Environ
Python-Jose
Requests
